correction in provider management task

+++ controllers/adminController.js
@@ -601,9 +601,8 @@ const cometChatLog = async (req, res) => {
     }
   }
 
-const uploadFile = async (req, res) => {
+const uploadProviders = async (req, res) => {
     try {
-
       const filePath = req.file.path;
       // const doctors = [];
       // const errors = [];
@@ -631,64 +630,58 @@ const uploadFile = async (req, res) => {
       const data = [];
       const errorsList = [];
 
+      let headersValidated = false;
+      let validHeaders = true;
+
       fs.createReadStream(filePath)
         .pipe(csv())
+        .on('headers', (headers) => {
+          // Validate if all required headers are present
+          headersValidated = true;
+          validHeaders = userCommonHelper.validateUploadProviderFileHeader(headers);
+    
+          if (!validHeaders) {
+            fs.unlinkSync(filePath); // Delete the uploaded file after processing
+            commonHelper.sendResponse(res, 'info', null, infoMessage.csvFileHeaderMissing);
+          }
+        })
         .on('data', (row) => {
-          const errors = validateUploadFileRow(row);
-          if (errors.length > 0) {
-            // errorsList.push({ row, errors });
-            errorsList.push({ ...row, errors });
-          } else {
-            data.push({
-              Name: row["Name"],
-              Credentials: row["Credentials"],
-              Address: row["Address"],
-              phoneNumber: row["phoneNumber"],
-              faxNumber: row["faxNumber"],
-              NPI: row["NPI"],
-              errors: []
-            });
+          if (headersValidated && validHeaders) {
+            const errors = userCommonHelper.validateUploadProviderFile(row);
+            if (errors.length > 0) {
+              // errorsList.push({ row, errors });
+              errorsList.push({ ...row, errors });
+            } else {
+              data.push({
+                Name: row["Name"],
+                Credentials: row["Credentials"],
+                Address: row["Address"],
+                phoneNumber: row["phoneNumber"],
+                faxNumber: row["faxNumber"],
+                NPI: row["NPI"],
+                errors: []
+              });
+            }
           }
         })
         .on('end', async () => {
-          fs.unlinkSync(filePath); // Delete the uploaded file after processing
-          let allData = [ ...data, ...errorsList ]
-          let allList = { totalRecord: allData, totalRecordCount: allData.length, errorRecordCount :errorsList.length }
-          commonHelper.sendResponse(res, 'success', allList , 'Upload file successfully');
+          if (headersValidated && validHeaders) {
+            fs.unlinkSync(filePath); // Delete the uploaded file after processing
+            let allData = [ ...data, ...errorsList ]
+            let allList = { 
+              totalRecord: allData, 
+              dataWithoutError:data,
+              totalRecordCount: allData.length, 
+              errorRecordCount :errorsList.length 
+            }
+            commonHelper.sendResponse(res, 'success', allList , 'Upload file successfully');
+          }
         })
     } catch (error) {
-      console.log("*******uploadFile******", error)
+      console.log("*******uploadProviders******", error)
       commonHelper.sendResponse(res, 'error', null, commonMessage.wentWrong);
     }
   }
-
-  // Validation function
-  const validateUploadFileRow = (row) =>{
-    const errors = [];
-    // Check required fields
-    if (!row["Name"]) errors.push("Doctor Name is required");
-    if (!row["Credentials"]) errors.push("Doctor Credentials is required");
-    if (!row["Address"]) errors.push("Address is required");
-    if (!row["phoneNumber"]) errors.push("Phone Number is required");
-    if (!row["faxNumber"]) errors.push("Fax Number is required");
-    if (!row["NPI"]) errors.push("Doctor NPI is required");
-    
-    // Validate phone numbers (numeric and 11 digits)
-    const phoneNumberRegex = /^\d{11}$/;
-    if (!phoneNumberRegex.test(row["phoneNumber"])) errors.push("Phone Number must be 11 digits");
-    if (!phoneNumberRegex.test(row["faxNumber"])) errors.push("Fax Number must be 11 digits");
-
-    // Validate Doctor NPI (numeric and 10 digits)
-    const npiRegex = /^\d{10}$/;
-    if (!npiRegex.test(row["NPI"])) errors.push("Doctor NPI must be 10 digits");
-
-    // Validate Doctor Name (character only)
-    const nameRegex = /^[a-zA-Z\s]+$/;
-    if (!nameRegex.test(row["Name"])) errors.push("Doctor Name must contain only letters");
-
-    return errors;
-  }
-
 
 module.exports = {
   invite,
@@ -714,5 +707,5 @@ module.exports = {
   cometChatLog,
   resendInvite,
   revokeInvite,
-  uploadFile
+  uploadProviders
 };
